Code Monitors: save result payload in db (#30056)
This modifies code monitor query runs to save the set of results it
receives from the API in the database.

Why do that? Because repo-aware code monitors will be very
difficult to re-run exactly until we have a better intermediate
representation for our search queries, so being able to show the results
in the UI means we need to save the results when we get them.
Additionally, this makes it much easier to audit and debug.

Because code monitor runs are only saved in the db for one day, there
is a bound on the size of additional information being stored in the db,
so I'm not too concerned about storage bloat.