Code Monitors: propagate CodeMonitorID through search tree (#30371)
This adds a new GraphQL endpoint that wraps the `search()` endpoint and propagates the new CodeMonitorID parameter through the search stack down to the CommitSearch job. This is all that is needed in the search pipeline -- remaining changes will be isolated to the CommitSearch job and the code monitors packages.

Additionally, I expect all of this change to be temporary. Once we can generate a job tree without depending on resolvers, I'd like to generate the tree, mutate the tree, and run the tree all in the worker job.