inline Aggregator.DoSearch
The only thing this method was doing was wrapping the job with a trace
(which was already happening in DoResults) and aggregating the error,
which is useless since it is only ever possible for a single error to be
aggregated.

My goal here is just to pull some threads to see how far I can get this
to unravel.