pass an aggregating stream to searchResolver.Stats()
This makes the call to `doResults` in `searchResolver.Stats()` take a
aggregating stream. Since `Stats()` is only called by reflection over
the GraphQL API, we know that `r.stream` will always be nil, so we are
not overwriting a stream that might be forwarding the results.
Additionally, this drops any alerts that might be returned by
`doResults`. They were not being used anyways, so this is no change in
behavior. I'm hoping to move alerts out of `doResults` anyways, so this
will be even less confusing in the near future.