pass an aggregating stream to doResults in searchResultsStats.getResults
This passes an aggregating stream to the `doResults` call in
`searchResultsStats.getResults`. Previously, this would pass in the
embedded `searchResolver`'s stream, which was always `nil` because
`getResults` is only called by `Languages` which is only used by the
GraphQL resolver. In addition, this gets rid of an unnecessary
wrap/unwrap of the results in a `SearchResultsResolver`.