internal: Add comments for generated metric names (#30473)
When debugging an issue related to these metrics, we tend to do a
string search for the metric name but will fail to find a match since
these metrics are generated. As a result, adding a comment will help
with such scenarios.

Also, in this commit a metric for bitbucket cloud is renamed to remove
a redundant string from the name.