Convert `gitserver.DefaultClient` to an interface (#30443)
This creates a gitserver.Client interface, and converts gitserver.DefaultClient to be that type. The motivation here is that I really want to be able to swap out DefaultGitserver with a mocked version in search tests, but that's currently not possible. Ideally, we'd go as far as to pass in a gitserver.Client object to the search code, but we're not there yet.