add a minimal amount of tracing to search jobs
Search jobs are the best representation of our structure we have at this
point, but they are not represented in the traces. This adds some
minimal tracing to the top of each of the search jobs so they show up in
traces in the same structure they are defined. Also will be useful to
help debug a weird issue we're seeing in the k8s deployment.