always stream from evaluate (#30576)
This modifies the single place we call `evaluate` to always pass in a
non-nil stream. In order for this behavior to be correct, we need to
support a non-nil stream in `evaluateAnd` and `evaluateOr`. Currently,
streaming is disabled by `IsStreamingCompatible` whenever the query has
an and/or node, but this modifies the logic to only disable streaming
when there are multiple basic queries to be run.