Search backend: simplify returns (#30584)
Now that we know that r.evaluate() is always called with a non-nil stream, we can depend on the stream to handle capturing the results and the stats, so we don't need to return them in addition to sending them on the stream. This modifies evaluateJob, evaluateAnd, evaluateOr, evaluatePatternExpression, and evaluate to only return a *search.Alert and an error.