go get -u -t ./... (#30628)
Ran on all sub gomodules as well

  find . -name go.mod -exec dirname '{}' \; | while read -r d; do
    pushd $d;
    go get -u -t ./...;
    go mod tidy;
    popd;
  done

Additionally had to make a code change to the extsv/auth pkg.
jwt.StandardClaims is deprecated and suggests using this struct instead.
Using this struct means we no longer pass in a unix timestamp, but
instead a timestamp controlled by the jwt package. This should have the
same behaviour.