use smaller number of parallel senders (#30708)
I think this is causing failures in CI because we can't send 1000
messages in 100 milliseconds on CI. I expect this is just because CI is
CPU-loaded. This just runs it with a smaller number of parallel workers
so that it will hopefully fit in the batch time-window in a stable
manner.