remove lower-level limits when and/or jobs are enabled
This removes the lower-level limit and timeout wrapper jobs when and/or
jobs are enabled. With and/or jobs, these limits are set in
toEvaluateJob. Additionally, these limits make and jobs find fewer
results than it should since it limits each of the and subqueries rather
than just relying on the limit on number of results returned.