dev/ci: add RunType for nightly release branch healthchecks (#30581)
Introduces a RunType that will trigger builds on the $major.$minor and $major.$minor-1 branches corresponding to the minimum upgradeable version. The goal of this is to allow the Delivery and DevX team to keep tabs on the releaseability of recent release branches to enable patches to be shipped faster.

This RunType will be set up to be triggered via Buildkite scheduled builds once a day.