fix key equality for type CommitMatch
In a recent change, I added AuthorDate to the Key type so that we could
sort commit matches by date. However, the pointer type is not traversed
when comparing structs with struct equality, so equivalent commits from
two different sources were not being evaluated as the same commit. This
broke deduplication for and/or queries.