symbols: improvements to ctags build script (#30832)
The main improvement is we no longer ship the ctags build dependencies
in the symbols docker image. Otherwise this can be reviewed commit by
commit:

- remove seccomp support in ctags
- use all your cores to build ctags
- cleanup via trap for ctags-install-alpine

The cleanup of this script was motivated in wanting to share it with how
we build ctags in zoekt.

Test Plan: built and ran the docker image locally. Verified that
universal-ctags --help didn't fail. Pushed this branch to backend
integration.