search: Add ability to remove all entries from search stack (#30752)
This commit adds the ability to remove all entries from the search
stack. While introducing the delete button I also updated the "create
notebook" button and added tests for the deletion functions.

I also changed the logic so that the popover is always shown even if
there are no search entries. Without it, removing all (or the last)
entry causes the popover to disappear which is confusing.

I also set the width of the popover to a fixed size, I found it weird
when the distance between the "create notebook" and the "delete all"
buttons changed due to the size of the stack entries.