search: Show last added item on top of search stack (#30764)
This reverses the order in which search stack items are rendered so that
the most recent item is rendered on top. This only affects the UI. The
items are still stored in order so Notebook creation still works as
before.

Instead of introducing the `id` property I could have used the same
logic that is used to identify "duplicate" entries (for updating them),
but I think that this logic is going away anyway once entries are added
manually instead of automatically (in which case the action is
deliberate, so no duplicate detection is necessary).