search: Manually add items to search stack
This commit changes the behavior of the search stack to allow items
being added manually instead of automatically.

As a consequence this change also removes the detection of duplicate
items (the same file/search could be added multiple times if desired).
Whether or not this is disreable (or whether an item should be
highlighted somehow when it matches the current query/file) needs to be
discussed.

Instead of having the search stack component automatically detect
whether it's openend on a page that can be added to the stack, I'm
keeping the existing API (the pages "push" to the stack).
I decided against having the component detect the page because the only
way I could see that working is via inspect and parsing the URLs, and
that appeard too fragile to me (and also the URLs might not contain all
the information we want).
But if URL parsing is considered to be good enough then this could be
considered in the future.