cmd/gitserver: Guard decision to block command behind feature flag
The feature flag is turned off by default and allows us to flip it
when we are confident with the changes in production after observing
the logs. Also, it lets us flip it back to the default in case there
are any regressions.

This commit also adds test to validate existing behaviour with an
without the feature flag enabled.