gitserver: build p4-fusion as a seperate stage
When we introduced p4-fusion to our docker images it added a significant
amount of build time. In the case of server it doubled the build time.
This is a first step to improving that but only targetted at gitserver.
Next up we can do the same change to server, but I wanted to keep this
focussed.

The meat of this change is extracting the docker steps for p4-fusion
into a build script. This is in preparation for using it in server image
builds.

The main benefit is it is now a seperate change. Previously we ran into
linking issues but this was easy to resolve. You just need to install
libstdc++.

Note: I noticed some flaws in our build script. This commit doesn't fix
those it just does manual refactoring. Follow-up commits can improve the
process. The only real functional change it does is use a shell script
trap for cleanup.

Test Plan: A lot of docker building. Running the gitserver image and
checking that p4-fusion runs and "ldd p4-fusion" is happy.