codeintel: Fix repository inference for NPM packages.
Later code in gitserver.RepoInfo uses the repo name to infer a path
where the repo contents are cloned. NPM package have their contents
cloned under an npm/ subdirectory (similar to github.com/) under
.sourcegraph/repos. However, this code was just passing in the package
name instead of using the proper repo name which has an npm/ prefix,
and removes an @ from the scope (if a scope is present).