[cloud] Store expiry time for organization invitations
Return consistent URL for the invitation because expiry time on the
JWT is now the same as in the DB.

We also force an expiry of 7 days to all existing pending invitations
with a DB migration. This way, we will never have long-lasting invitations
anymore, that can be potentially misused years from now to implant
organization members by a potential attacker.