gitserver: pre-built p4-fusion artifacts
I was hoping all the improvements made around this script would result
in much faster build times. It does, but with a very high variance. As
such I've instead uploaded the p4-fusion binary to the
sourcegraph-artifacts bucket. I don't suspect we will often change
versions here so this seems like a good approach. Additionally the build
script is quite robust now, so when we do bump the version it will be
easy to build the new version.

I believe this complexity and cache is justified by the build times
introduced to our CI for this step.

Test Plan: Built image locally. Also pushing to main dry run branch to
exercise image building on CI.

  docker build -t p4-fusion --target=p4-fusion ."