ssbc: Prettify create batch change page (#30295)
Not meant to be pixel perfect but some start towards a more "wow" effect when getting started with SSBC the first time. The visibility toggle is muted and disabled, but it makes it visually more appealing.

Also fixes a bug where the namespace on the editor page would not match the namespace of the batch change being edited.