unembed SearchInputs from searchResolver
Because SearchInputs is embedded in searchResolver, it is difficult to
tell exactly where it is being used since find references doesn't show
the places its fields are accessed through the `searchResolver` type.

This just makes it a named field rather than an embedded field. This
will make it easier to reason about some of the upcoming changes I'd
like to make.