search: Refactor search stack UI widget (#31248)
This makes some updates to the look and feel of the widget. In
particular it changes the closed state of the widget from a small button
to showing the top item. Clicking on the item will open the widget.

The buttons for adding a new entry will now always be shown, even if the
entry was just added.

Annotations can now be added. Those are rendered as markdown blocks
*before* the references block.