codeintel: extend PackageDependency for dependency search
This commit extends the `reposource.PackageDependency` interface with
methods necessary for implementing dependency search (in a follow up PR).

Some methods of the added methods are already implemented by the
embedded type as long as it is a pointer, so we change signatures
everywhere to deal with pointers rather than values, which should have
the added benefit of avoiding a fair bit of unnecessary memory copies.

Part of #28932