search: Make search stack/notepad entries selectable
This commit adds the ability to select search stack/notepad entries via
click, ctrl/meta+click and shift+click.
The entries can also be selected via space, ctrl+space and shift+space
if they are focused.

In the end I went with role=listbox and role=option for the entries,
because I think that comes closes to how we want this component to
behave eventually. But note that more work is required to make this
component fully usable via the keyboard.