remove code monitor info in graphqlbackend
This removes some stuff that I had staged for repo-aware code monitors.
It is unused right now, and is just getting in the way. I tried to do
this once before, but removed the GraphQL endpoint that was still being
used, which broke things. This does not break things. I have tested it.