Search backend: extract results into new package `execute` (#31759)
This extracts the logic of searchResolver.results() into a new function execute.Execute. I created a new package because of circular imports. I think this would probably better belong in run, but the fact that we use SearchInputs in other dependent packages makes that not possible. We should be able to clean this up more easily once we have better partitioned.