batches: Default to oldest global external service
We had an issue in one of our instances where a new global external service was configured and then batch changes broke by that. If we said we default to the _oldest_ external service instead, this would never change unintentionally. It is sort of a breaking change, but one for a deprecated feature that never really was documented how it works. So my personal feeling is that this is alright.
While at it, I also removed user added external services, using those didn't make sense in the first place as long as we don't enforce ownership.