Search backend: remove stream from searchResolver (#31813)
This uses the new SearchClient for the compute streaming endpoint. 
Then, now that no streaming clients use searchResolver, we can remove 
stream from the search args and inline resultsBatch, removing resultsStream.
Finally, since we only care about whether we're streaming or not during 
planning phase (we just check if stream is nil or not), we can just pass a 
protocol into the Plan() function.