explicitly make gitserver a dependency of commit search
This moves the gitserver client onto the CommitSearch job as opposed to
implicitly using the global client inside the job. This allows us to
inject a mock gitserver client. As mentioned in a previous comment, I'd
like to move towards passing these dependencies in at execution time
rather than at planning time, but we currently follow this pattern for
zoekt and searcher, so I'm maintaining consistency for now.