search: Quickly sync new versions of dependency repos (#31925)
We recently introduced a way to lazy sync package repos via repo-updater's lookup endpoint.
This worked well for entirely new package repos, but not new versions of existing package repos,
since we don't enqueue a high priority update for existing repos.

This PR addresses that problem by introducing an Update argument to the
endpoint that allows us to request such an update. We could just issue
an additional request to the enqueue-repo-update endpoint, but this approach avoids the extra roundtrip.