Hide reference panel when not viewing code (#31849)
This fixes #31640 by only showing the experimental reference panel when
viewing files. Previously the panel would stay open when navigating from
a file view to a directory view, or from file view to search results.

In order to implement this I moved all of the state related to the
reference panel to the `RepoRevisionContainer` and also only render the
reference panel in there. The nice side-effect of that is that we now
don't re-render `Layout` on any change to the clicked token anymore.