codeintel locations: fix URLs by using git blob as resource
This fixes inconsistent behaviour in our `GitBlobLSIFData` resolver that
would cause the `Locations` being returned in the
`definitions`/`references`/... having `/tree/` as opposed to `/blob/` in
their URL.

`tree` is only meant to be used for directories and not files, though,
which is why the produced URLs that include line ranges would be
redirected by our web application.

So this changes the second parameter of `gql.CreateFileInfo` to
correctly say that the locations are *not* directories, but files.