Experimental ref panel: restore click2j2d behaviour (#31984)
This fixes #31642 by reverting to the old behaviour of the reference panel and only opening it when clicking on the token _at the definition_ site.

It does this by making the complete panel and its state driven by the URL and the `#tab=references` hash.

All of the other code related to the clicked token, the global props, etc. is gone.