repo-updater: Update out parameter on missed code path
The 'sourced' parameter is used as an "out parameter",
initially its ID is not set, but it will set when the repo
is created. However, on this code path, the ID was not updated.

Since the returned diff is ignored in syncRepo, it makes
sense to update the out parameter, otherwise the new ID is "lost".

This fixes a downstream crash when trying to enqueue an update
for a repo with ID == 0.