docs: Remove warning about Batch Changes renaming
I ran into this while doing something different in the docs and I have
to say: the first impression when seeing something orange with
"WARNING:" in it on the top of a docs page is not good.

So since it's been over 10 releases, I think it's fine to remove it.