docs: change structure of code intel docs start page
I've changed the structure of the start page of our docs to answer the
following questions above the fold:

* what does code intelligence give me
* what does code intelligence consist of
* how can I start using it
* what do I need to do to use which feature
* where can I use code intelligence

As always: this is the first iteration. I have some more ideas on how to
tweak this and improve it a bit more, but I think we should really start
answering these questions at the start of our docs, *especially* since
we link to our docs from the product.