authz/github: allow list group repos 404 to persist to cache (#32357)
Potential solution to sourcegraph/customer#742 . Looks like GitHub has some odd behaviours related to teams called owners etc. that may cause "list repo" on an existing repo to return 404. A 404 is not transient, and retrying on a 404 will not make it un-404, so we allow these errors to persist to cache. I've added a docstring inline to explain why this is safe.

Co-authored-by: Molly Weitzel <molly.weitzel@sourcegraph.com>