database: Remove database.Mocks.ExternalServices (#32437)
The ValidateConfig method was extracted from the ExternalServiceStore
and renamed to ValidateExternalServiceConfig.

Instead, it is now a function that takes the store as a parameter. This
allows us to mock the store functions using our new mocking system and
allows us to remove the global database.Mocks.externalServices.

Because we have custom validators when running in enterprise mode
the validation function is exposed as a variable so that we can swap
in different implementations. This is similar to we had before where
the store constructor was exposed as a variable.

Once all of the above refactoring was done, the external service mock
could be removed.