Backend: add AnonymousUID to Actor struct (#32475)
This adds an AnonymousUID field to actors. There are a few places
(feature flags, event logs) where we need to pass through an HTTP
request so that the anonymous UID can be pulled from the cookie.
Populating the AnonymousUID as part of the cookie auth step will allow
us to use the actor as the only source of information for "this is the
current user" throughout the application code.