search: Fix backspace item deletion bug
This was caused by the fact that I switched from listing to key *up*
events to key *down* events, but didn't make the same change to the
event handler bound to the textarea, so all key events bubbled up to the
parent.

Instead of relying on stopping the event propagation, the list event
handler now checks whether an annotation textarea is focused and if yes
it ignores the event.

This commit also adds the ability to remove focus from the annotation
textarea by pressing Escape.