break feature flag request dependency
This breaks the dependency between feature flags and a request object.
Now that AnonymousUID exists on the actor struct, we don't need to hold
on to the request object for the cookie and we can depend purely on the
actor for the information we need to populate feature flags.

Additionally, this fixes an issue where, if the actor is changed on a
context, the feature flags returned may still be for the first actor.