Repo aware monitors: add hook in CommitSearch job (#32576)
This adds a hook function to the CommitSearch job that will wrap the search call with logic specific to code monitors. Because repo resolution and revision resolution are currently very tied together, I can't just mutate the inputs to the CommitSearch job yet, so for the moment we're stuck with this hook function. Thankfully, the footprint is pretty minimal. All the rest of the logic will live in code monitor packages.