feat: Add syntax highlighting configuration (#32084)
With a new site config setting, it is now possible to configure what language
a file should be highlighted as via:
- extension
- pattern matching

This allows people to configure highlighting and add languages (and language
mappings) without having to send PRs to Sourcegraph. In the future, this may
be extended to be able to provide more configuration for highlighting as well.

Finally, we also add the ability to set defaults for the highlighting engine (as of
right now, that is syntect or tree-sitter) both for the site and for individual language
overrides. Follow up PRs will be sent to update more documentation and
information.