search: create and document git-stats script
This is a script we have shared directly with customers before to
understand the scale of the monorepos. This now stores it in our
repository and documents it under our monorepo documentation.

Test Plan: ran git-stats on the sourcegraph repo. Note the links in the
documentation will only work once this PR has landed.